rqe_iterators: factor out build_union into union_opaque
Move the union C-boundary construction (build_union and its
ProfileChildren callback) from iterators_ffi into
rqe_iterators::union_opaque, and route the FFI NewUnionIterator together
with the numeric and geo range iterators through it. This lets both the
FFI constructors and the upcoming query_eval dispatcher build the
identical wrapper shape. The callback now profiles children via
CRQEIterator::into_profiled directly.
